requirements.txt
pandas==2.0.0
ipymarkup==0.9.0
nerus==1.7.0
Flask==2.2.3
waitress==2.1.2
python-telegram-bot==20.2
PyYAML==6.0
requests==2.28.2
numpy==1.24.2
torch==1.13.0
wget==3.2
tqdm==4.65.0
scikit-learn==1.2.2